fix(MRTD): ensure MRTD PoP challenge is skipped when `challenge_info` is absent
